requirements.txt
flask
torch
torchvision
pillow
firebase-admin
flask-cors
